Fin de la fonction d'ajout des clients, Activation/desactication, Suppression sans notification des utilisateurs

+++ application/modules/compagnie/views_compagnie.py
@@ -1,12 +1,9 @@
 __author__ = 'User'
 
-
 from ...modules import *
 from forms_compagnie import FormClient
-from models_compagnie import Compagnie
-from ..site.models_site import Secteur
-from ..user.forms_user import FormUser
-
+from models_compagnie import Compagnie, Categorie
+from ..user.models_user import Users
 
 prefix = Blueprint('client', __name__)
 
@@ -14,7 +11,6 @@ prefix = Blueprint('client', __name__)
 @prefix.route('/')
 @login_required
 def index():
-
     title_page = 'Clients'
 
     datas = Compagnie.objects()
@@ -25,17 +21,26 @@ def index():
 @prefix.route('/view/<objectid:client_id>')
 @login_required
 def view(client_id):
-
     title_page = 'Clients'
 
-    data = Client.objects().get(id=client_id)
+    data = Compagnie.objects.get(id=client_id)
     form = FormClient(obj=data)
-    form.secteur_id.data = str(data.secteur_id.id)
+    form.id.data = client_id
 
-    all = Secteur.objects(actif=True)
-    form.secteur_id.choices = [('', ' ')]
+    form.idcategorie.data = [str(cat.id) for cat in data.idcategorie]
+
+    form.maincategorie.data = str(data.maincategorie.id)
+
+    form.maincategorie.choices = [('', 'Faite le choix de la categorie principale')]
+
+    for choice in data.idcategorie:
+        form.maincategorie.choices.append((str(choice.id), choice.name))
+
+    all = Categorie.objects(Q(activated=True) & Q(parent_idcategorie__ne=None))
+
+    form.idcategorie.choices = [('', '')]
     for choice in all:
-        form.secteur_id.choices.append((str(choice.id), choice.name))
+        form.idcategorie.choices.append((str(choice.id), choice.name))
 
     return render_template('client/view.html', **locals())
 
@@ -45,42 +50,91 @@ def view(client_id):
 @login_required
 @roles_required([('super_admin', 'client')], ['edit'])
 def edit(client_id=None):
-
     title_page = 'Clients'
 
     if client_id:
-        data = Client.objects.get(id=client_id)
+        data = Compagnie.objects.get(id=client_id)
         form = FormClient(obj=data)
         form.id.data = client_id
-        form.secteur_id.data = str(data.secteur_id.id)
+
+        form.idcategorie.data = [str(cat.id) for cat in data.idcategorie]
+
+        form.maincategorie.choices = [('', 'Faite le choix de la categorie principale')]
+
+        if request.method == 'GET':
+
+            form.maincategorie.data = str(data.maincategorie.id)
+
+            for choice in data.idcategorie:
+                form.maincategorie.choices.append((str(choice.id), choice.name))
     else:
-        data = Client()
+        data = Compagnie()
         form = FormClient()
 
-    all = Secteur.objects(actif=True)
-    form.secteur_id.choices = [('', ' ')]
+    all = Categorie.objects(Q(activated=True) & Q(parent_idcategorie__ne=None))
+
+    form.idcategorie.choices = [('', '')]
     for choice in all:
-        form.secteur_id.choices.append((str(choice.id), choice.name))
+        form.idcategorie.choices.append((str(choice.id), choice.name))
+
+    if not client_id:
+        form.maincategorie.choices = [('', 'Faite le choix de la categorie principale')]
+
+    for idchoice in request.form.getlist('idcategorie'):
+        choice = Categorie.objects.get(id=idchoice)
+        form.maincategorie.choices.append((str(choice.id), choice.name))
 
     if form.validate_on_submit():
 
         data.raison = form.raison.data
         data.name = form.name.data
-        data.email = form.email.data
-        data.phone = form.phone.data
         data.ville = form.ville.data
         data.quartier = form.quartier.data
-        data.urlsite = form.urlsite.data
-        data.client = True
-        data.fournisseur = False
-        data.adress = form.adress.data
-        data.bp = form.bp.data
-        data.rue = form.rue.data
+        data.adresse = form.adresse.data
+        data.postal_code = form.postal_code.data
+        data.repere = form.repere.data
 
-        secteur = Secteur.objects.get(id=form.secteur_id.data)
-        data.secteur_id = secteur
+        data.email = form.email.data
+        data.phone = form.phone.data
+        data.description = form.description.data
+        data.urlsite = form.urlsite.data
+
+        data.latitude = form.latitude.data
+        data.longitude = form.longitude.data
+
+        data.facebook = form.facebook.data
+        data.twitter = form.twitter.data
+        data.linkedin = form.linkedin.data
+        data.youtube = form.youtube.data
+
+        old_categorie = data.idcategorie
+        data.idcategorie = []
+
+        for cat_id in request.form.getlist('idcategorie'):
+
+            the_cat = Categorie.objects.get(id=cat_id)
+
+            data.idcategorie.append(the_cat)
+
+        principale = Categorie.objects.get(id=form.maincategorie.data)
+        data.maincategorie = principale
 
         data = data.save()
+
+        ## verifier que les nouvelles categories ont les informations de l'entreprise cours
+        for cat_id in request.form.getlist('idcategorie'):
+
+            the_cat = Categorie.objects.get(id=cat_id)
+
+            if data not in the_cat.idcomapagnie:
+                the_cat.idcomapagnie.append(data)
+                the_cat.save()
+
+        ## Enlever les informations de l'entreprise dans les anciennes categories qui ne l'appartienne
+        for old_cat in old_categorie:
+            if old_cat not in data.idcategorie:
+                old_c = Categorie.objects(id=old_cat.id)
+                old_c.update_one(pull__idcomapagnie=data)
 
         flash('Enregistement effectue avec succes', 'success')
         return redirect(url_for('client.view', client_id=data.id))
@@ -88,12 +142,10 @@ def edit(client_id=None):
     return render_template('client/edit.html', **locals())
 
 
-
 @prefix.route('/deleted', methods=['POST'])
 @login_required
 @roles_required([('super_admin', 'client')], ['delete'])
 def deleted():
-
     from ..opportunite.models_opportunite import Opportunite
     from ..document.models_doc import Document
 
@@ -102,19 +154,26 @@ def deleted():
     count = 0
     for item in request.form.getlist('item_id'):
         info = {}
-        item_found = Client.objects().get(id=item)
+        item_found = Compagnie.objects().get(id=item)
         opportunite = Opportunite.objects(client_id=item_found)
 
         if opportunite:
             info['statut'] = 'NOK'
-            info['message'] = 'Le Client "'+item_found.name+'" est utilise par '+str(opportunite.count())+' autre(s) opportunite(s)'
+            info['message'] = 'Le Client "' + item_found.name + '" est utilise par ' + str(
+                opportunite.count()) + ' autre(s) opportunite(s)'
 
         exit_suivie = Document.objects(client_id=item_found)
         if exit_suivie:
             info['statut'] = 'NOK'
-            info['message'] = 'Le Client "'+item_found.name+'" est utilise par '+str(exit_suivie.count())+' autre(s) Documents(s)'
+            info['message'] = 'Le Client "' + item_found.name + '" est utilise par ' + str(
+                exit_suivie.count()) + ' autre(s) Documents(s)'
 
-        if not opportunite and not exit_suivie:
+        if not item_found.uploaded:
+            info['statut'] = 'NOK'
+            info['message'] = 'Le Client "' + item_found.name + '" ne peut pas etre supprime car il n\'a pas ete ajoute' \
+                                                                'manuellement '
+
+        if not opportunite and not exit_suivie and item_found.uploaded:
             item_found.delete()
             element.append(str(item_found.id))
             count += 1
@@ -124,7 +183,7 @@ def deleted():
     if count:
         info = {}
         info['statut'] = 'OK'
-        info['message'] = str(count)+' element(s) ont ete supprime avec success'
+        info['message'] = str(count) + ' element(s) ont ete supprime avec success'
         info['element'] = element
         data.append(info)
 
@@ -133,78 +192,113 @@ def deleted():
     return data
 
 
-@prefix.route('/contact/view/<objectid:contact_id>')
-@login_required
-@roles_required([('super_admin', 'client')])
-def contact_view(contact_id):
+@prefix.route('/reload/categorie', methods=['POST'])
+def reload_categorie():
+    list = []
 
-    title_page = 'Clients'
+    idcategorie = request.json['data']
 
-    data = Users.objects.get(id=contact_id)
-    form = FormUser(obj=data)
-    form.id.data = str(data.id)
-    if data.civilite:
-        form.civilite.data = str(data.civilite.id)
+    for data in idcategorie:
+        current_cat = Categorie.objects.get(id=data)
+        current = {}
+        current['id'] = str(current_cat.id)
+        current['name'] = current_cat.name
+        list.append(current)
 
-    all = Civilite.objects()
-    form.civilite.choices = [('', 'Choisir la civilite ')]
-    for choice in all:
-        form.civilite.choices.append((str(choice.id), choice.name))
+    datas = json.dumps({
+        'statut': 'OK',
+        'data': list
+    })
 
-    return render_template('contact/view.html', **locals())
+    return datas
 
 
-@prefix.route('/contact/edit/<objectid:client_id>/<objectid:contact_id>', methods=['GET', 'POST'])
-@prefix.route('/contact/edit/<objectid:client_id>', methods=['GET', 'POST'])
+@prefix.route('/activated', methods=['POST'])
 @login_required
 @roles_required([('super_admin', 'client')], ['edit'])
-def contact_edit(client_id, contact_id=None):
+def etat_activated():
 
-    title_page = 'Clients'
+    data = []
+    element = []
+    count = 0
+    for item in request.form.getlist('item_id'):
+        info = {}
+        item_found = Compagnie.objects().get(id=item)
 
-    if contact_id:
-        data = Users.objects.get(id=contact_id)
-        form = FormUser(obj=data)
-        form.id.data = str(data.id)
-        if data.civilite:
-            form.civilite.data = str(data.civilite.id)
+        if item_found.activated:
+            info['statut'] = 'NOK'
+            info['message'] = 'L\'entreprise "'+item_found.name+'" est deja active'
+
+        if not item_found.activated:
+            item_found.activated = True
+            element.append(str(item_found.id))
+            item_found.save()
+            count += 1
+        else:
+            data.append(info)
+
+    if count:
+        info = {}
+        info['statut'] = 'OK'
+        info['message'] = str(count)+' element(s) ont ete active avec success'
+        info['element'] = element
+        data.append(info)
+
+    data = json.dumps(data)
+
+    return data
+
+
+@prefix.route('/unactivated', methods=['POST'])
+@login_required
+@roles_required([('super_admin', 'client')], ['edit'])
+def etat_unactivated():
+
+    data = []
+    element = []
+    count = 0
+    for item in request.form.getlist('item_id'):
+        info = {}
+        item_found = Compagnie.objects().get(id=item)
+
+        if not item_found.activated:
+            info['statut'] = 'NOK'
+            info['message'] = 'L\'entreprise "'+item_found.name+'" est deja desactive'
+
+        if item_found.activated:
+            item_found.activated = False
+            element.append(str(item_found.id))
+            item_found.save()
+            count += 1
+        else:
+            data.append(info)
+
+    if count:
+        info = {}
+        info['statut'] = 'OK'
+        info['message'] = str(count)+' element(s) ont ete desactive avec success'
+        info['element'] = element
+        data.append(info)
+
+    data = json.dumps(data)
+
+    return data
+
+
+@prefix.route('/etat/<objectid:client_id>')
+@login_required
+@roles_required([('super_admin', 'client')], ['edit'])
+def etat(client_id):
+    client = Compagnie.objects.get(id=client_id)
+    if client.activated:
+        client.activated = False
     else:
-        data = Users()
-        form = FormUser()
+        client.activated = True
 
-    all = Civilite.objects()
-    form.civilite.choices = [('', ' Choisir la civilite ')]
-    for choice in all:
-        form.civilite.choices.append((str(choice.id), choice.name))
+    client.save()
 
-    success = False
-    if form.validate_on_submit():
-
-        data.first_name = form.first_name.data
-        data.last_name = form.last_name.data
-        data.email = form.email.data
-
-        if form.civilite.data:
-            civilite = Civilite.objects(id=form.civilite.data).first()
-            data.civilite = civilite
-
-        data.fonction = form.fonction.data
-        data.mobile = form.mobile.data
-        data.telephone = form.telephone.data
-        data.note = form.note.data
-
-        data = data.save()
-
-        if not contact_id:
-            current_client = Client.objects.get(id=client_id)
-            contact_compagnie = CompagniUser()
-            contact_compagnie.user_id = data
-            current_client.contact.append(contact_compagnie)
-            current_client.save()
-
-        success = True
-
-    return render_template('contact/edit.html', **locals())
+    flash('Les modifications de l\'etat du client ont ete effectue', 'success')
+    return redirect(url_for('client.edit', client_id=client_id))
 
 
 # @prefix.route('/find/customer/<objectid:customer_id>', methods=['GET','POST'])
@@ -212,7 +306,7 @@ def contact_edit(client_id, contact_id=None):
 def find_customer():
     customer_id = str(request.json['client'])
 
-    customer = Client.objects.get(id=customer_id)
+    customer = Compagnie.objects.get(id=customer_id)
 
     data = json.dumps({
         'id': str(customer.id),
@@ -231,7 +325,7 @@ def find_customer():
 def find_contact():
     customer_id = str(request.json['client'])
 
-    customer = Client.objects.get(id=customer_id)
+    customer = Compagnie.objects.get(id=customer_id)
     customer = customer.contact
 
     list = []
@@ -244,12 +338,11 @@ def find_contact():
         current['last_name'] = data.user_id.last_name
         current['fonction'] = None
         if data.user_id.fonction:
-            current['fonction'] = '('+data.user_id.fonction+')'
+            current['fonction'] = '(' + data.user_id.fonction + ')'
         # current['email'] = data.user_id.email
         # current['mobile'] = data.user_id.mobile
         # current['telephone'] = data.user_id.telephone
         list.append(current)
-
 
     datas = json.dumps({
         'statut': 'OK',
@@ -261,7 +354,6 @@ def find_contact():
 
 @prefix.route('/find/contact/', methods=['POST'])
 def find_single_contact():
-
     customer_id = str(request.json['contact'])
 
     customer = Users.objects.get(id=customer_id)
@@ -285,64 +377,7 @@ def find_single_contact():
     return data
 
 
-@prefix.route('/contact/deleted', methods=['POST'])
-@login_required
-@roles_required([('super_admin', 'client')], ['delete'])
-def contact_deleted():
-
-    from ..opportunite.models_opportunite import Opportunite
-    from ..document.models_doc import Document
-    from ..user.models_user import Users
-
-    data = []
-    element = []
-    count = 0
-    for item in request.form.getlist('item_id'):
-        info = {}
-        item_found = Users.objects().get(id=item)
-        opportunite = Opportunite.objects(contact_id=item_found)
-
-        if opportunite:
-            info['statut'] = 'NOK'
-            info['message'] = 'Le contact "'+item_found.first_name+' '+item_found.last_name+'" est utilise par '+str(opportunite.count())+' autre(s) opportunite(s)'
-
-        exit_suivie = Document.objects(contact_id=item_found)
-        if exit_suivie:
-            info['statut'] = 'NOK'
-            info['message'] = 'Le contact "'+item_found.first_name+' '+item_found.last_name+'" est utilise par '+str(exit_suivie.count())+' autre(s) Documents(s)'
 
 
-        if not opportunite and not exit_suivie:
-            item_found.delete()
-            element.append(str(item_found.id))
-            count += 1
-        else:
-            data.append(info)
 
-    if count:
-        info = {}
-        info['statut'] = 'OK'
-        info['message'] = str(count)+' element(s) ont ete supprime avec success'
-        info['element'] = element
-        info['reload'] = 1
-        data.append(info)
 
-    data = json.dumps(data)
-
-    return data
-
-@prefix.route('/etat/<objectid:client_id>')
-@login_required
-@roles_required([('super_admin', 'client')], ['edit'])
-def etat(client_id):
-
-    client = Client.objects.get(id=client_id)
-    if client.actif:
-        client.actif = False
-    else:
-        client.actif = True
-
-    client.save()
-
-    flash('Les modifications de l\'etat du client ont ete effectue', 'success')
-    return redirect(url_for('client.edit', client_id=client_id))
